+++ Tests/TestConsole/SpaceShip.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -22,34 +21,6 @@ namespace TestConsole
         {
             _Position = Position;
 
-        }
-    }
-
-    struct Vector2D
-    {
-        private double _X;
-        //private double _Y;
-
-        public double X
-        {
-            get
-            {
-                return _X;
-            }
-            set
-            {
-                _X = value;
-            }
-        }
-        //public double Y { get => _Y; set => _Y = value; }
-        public double Y { get; set; }
-
-        public double Length => Math.Sqrt(X * X + Y * Y);
-
-        public Vector2D(double x, double y)
-        {
-            this._X = x;
-            this.Y = y;
         }
     }
